moving unused functions
